moved controler function to controller file
